Move Toolbar to ui folder
