+++ frontend/svelte/src/routes/Neurons.svelte
@@ -4,7 +4,7 @@
   import type { Unsubscriber } from "svelte/types/runtime/store";
   import { AuthStore, authStore } from "../lib/stores/auth.store";
   import { i18n } from "../lib/stores/i18n";
-  import Toolbar from "../lib/components/Toolbar.svelte";
+  import Toolbar from "../lib/components/ui/Toolbar.svelte";
   import NeuronCard from "../lib/components/neurons/NeuronCard.svelte";
 
   // TODO: To be removed once this page has been implemented
